Rename company details controller
Previously the name of the details controller did not match the template. Changed the company details controller name to be correct.
