+++ src/apps/companies/controllers/details.js
